refactor: improve code architecture following PSR-12 and SRP principles
- Apply Dependency Injection pattern across all components
- Optimize Loader class to handle component initialization
- Implement view templates to separate business logic from presentation
- Split ApiSettings class into smaller, specialized components (Model, Service, Controller)
