+++ src/Admin/GraduateColumnsManager.php
@@ -44,4 +44,3 @@ class GraduateColumnsManager
         }
     }
 }
-
